% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootrlm.R
\name{bootrlm}
\alias{bootrlm}
\title{Fit bootstrapped robust linear models}
\usage{
bootrlm(formula, data, r, method, options)
}
\arguments{
\item{formula}{A formula describing the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{r}{An integer number of replicates.}

\item{method}{A character string specifying the estimation method.}

\item{options}{A named list with options for particular estimation methods.}
}
\value{
An object of class \code{bootrlm}, containing the following
components:
\item{r}{The integer number of replicates.}
\item{replicates}{A matrix whose columns are the row indices of each
replicate.}
\item{coefficients}{A matrix whose columns are the parameter
estimates for each replicate.}
\item{fitted}{A matrix whose columns are the fitted values
corresponding to the matrix \code{replicates}.}
\item{residuals}{A matrix whose columns are the residuals
corresponding to the matrix \code{replicates}.}
\item{scale}{A vector with the estimated scale for each replicate.}
\item{rng_state}{The random number generator state.}
\item{formula}{The formula supplied.}
\item{data}{The data frame supplied.}
\item{method}{The estimation method requested.}
}
\description{
Fit bootstrapped robust linear models
}
\examples{
data(stackloss)
bootrlm_fit <- bootrlm(stack.loss ~ ., stackloss, r = 1000, method = "MM")
}

